Add support for command-line arguments
We now allow CLI arguments to be passed to the CP/M binaries that
we execute.  To prove that works I've written a simple Z80 binary
that shows its own arguments, and tested it under real hardware
and another emulator.

Use "make" to build the samples, with pasmo.

Sample run:

    $ ./go-cpm samples/cli-args.com i like cake
    The command-line argument supplied was 'i like cake'

This closes #1.