Allow multiple open files
This pull-request closes #27, by updating our file-handling to
allow multiple open files at once.  Instead of storing a single
file-handle when a file was created/opened, we installed stash
the filehandle of the open-files we've created in the FCB.

We (ab)use the Al[0] field to store our file number and retrieve
it later.

Testing:

* Loading zork works
  * Which uses random-reads to load zork.dat
* Loading/Saving games in zork works.

More than that?  Who knows.