Allow turbo-pascal-compiled binaries to run
This pull-request closes #60, by allowing binaries compiled by
turbo-pascal to execute - rather than aborting with "out of memory"
errors.

It turns out the generated binaries look at the contents of low-ram
to see how much RAM is free, by looking for the address of BDOS in
a jump instruction.  If that was unset they believed there was no
free RAM.