Resolved confusion about BDOS vs BIOS
This pull-request closes #68 by resolving the confusion about BIOS
and BDOS - I understand now why I had issues with mixing RawIO and
mixed IO it was because I was using the wrong handler for the same
things!

This pull-request:

* Explicitly marks the majority of our syscalls as being BDOS functions.
* Implements the four necessary BIOS calls to make my programs work.

This completely resolves the issue I had with laggy LIST and PRINT in
Microsoft BASIC, allows WORDSTAR.COM to launch, and fixes some other
oddities.

I'm so happy!