Use a stateful console-reading package
This pull-request changes the IO interface we're using to maintaining
state - although this doesn't (yet) remove the use of `stty` which
is tracked in #65, it does close #78 by avoiding the need to re-exec
the program after every character.

Keeping track of the current "echo" vs "noecho" state means that
we need to make this a global object, which we can reuse so we
do that here.