Allow Ctrl-C to reboot, at the CCP (#83)
This pull-request close #82, by changing our ReadLine implementation
to return a magic error if the user presses Ctrl-C twice in a row.

This then triggers a restart of the CCP.

This only works in the CCP, and in programs it launched.  If a user
presses this combination otherwise it will terminate the emulation.

In a real CP/M system  Ctrl-C would reload, but it is too easy to
accidentally press, so I ensure it is two consecuitive Ctrl-Cs that
triggers this behaviour.