Always use drive-remapping (internally)
Rather than conditionally using subdirectories as drives, we now
just do that always.  If drives are in use then we have amapping:

* A: -> A/
* B: -> B/
etc

If they are not?

*A: -> .
*B: -> .

That means we can always use our lookup table (i.e. map) to know
where to look for things and don't need to have special cases.

This closes #101.