Implement a naive virtual filesystem
This pull-request will update #104 by adding support for a simple
virtual filesystem.  This means that when the user runs "DIR A"
they will see the files they should see and _also_ files from
static/A/.

The intention is that we can bundle some fixed files within the
emulator which will be useful.  Of course we're not implementing
full functionality so any binary that needs to read an associated
data-file, or similar, will fail.

We've just added support to the three BDOS calls required to make
the CCP work:

* Open
* Read
* Close

Anything beyond that will fail and that's okay.  But it does mean
we can bundle our files appropriately.