Added "#.com", which allows us to have comments
As everything in the CP/M CCP is a command we
can implement comments by adding a binary named
"#".

That means this "does nothing":

        # foo bar

What actually happens, of course, is that "#.com"
is loaded with "foo bar" passed as arguments to it.
But since the binary immediately terminates we get
a comment in practice.