Allow polling for input, correctly
This pull-request DROPS SUPPORT FOR WINDOWS, but adds the ability
to detect when there is pending console input ready to be read.

This closes #117.

I've tested on Linux, I think the other platforms are _probably_
okay.  But who knows?

Feedback most welcome.  I guess it will come if it affects people
if this is broken then playing zork will fail.  If this works we've
got access to more console-based games.