Implement support for the RunCPM-Specific function F_UPTIME
This pull-request implements BDOS call 0xF8 / 248, which is
the F_UPTIME syscall, designed to show elapsed time since startup.

We return the time in nanonseconds rather than milliseconds, just
to mix things up a little.  Though in practice the value of either
measurement will wrap around so quickly the increase granualarity is
almost worthless.

This PR doesn't add any of the other RunCPM-specific syscalls as
those have not yet been seen in the wild.  It is only the v5
of BBC BASIC which seems to call the F_UPTIME syscall for some reason.

This closes #158.