Avoid the unnecessary invokation of STTY
CP/M has a couple of BIOS functions for reading input from the user,
there is a function for reading a string and a number of functions
relating to reading a single character.  The character-reading functions
are in two types:

* Those that should _echo_ the input, as received.
* Those that should _not_ echo the input.

In the past I didn't want to deal with the per-system handling to get
this echo on/echo off support working in a good way.  Instead I just
executed the system `stty` binary to disable/enable echoing, and then
read STDIN in the naive way.

Unfortunately using `stty` this way was both slow, and non-portable.
Slow because I exected `stty` at every time a keystroke was needing to
be read, and non-portable because Microsoft Windows wouldn't work.

I solved the slowness problem by keeping track of the echo/no-echo state
to cut down on executing the binary more than the minimum number of
times required (i.e. only when the state needed to change).  But the
portability was still a concern.

This pull-request takes inspiration from the way that we support
multiple methods for console _output_ by introducing that same
flexibility for input.  We now have a generalized layer to select
between different input-handlers, and to make that worthwhile we've got
two of them:

* The old `stty`-based approach.
* A new approach based on the termbox-go library.

The intention is to test the termbox-based solution, and if it works as
expected it can become the default, and if not we can revert back to the
`stty`-based solution, which while inelegant has the advantage that it
does actually work.

This will close #65, once complete, but it needs testing on a MacOS host
and ideally Windows too.