Shrink binaries by a byte
The recent pull-request used a tail-call optimization, but used an
absolute jump (three bytes) instead of a relative one (two bytes):

* https://github.com/skx/cpmulator/pull/168

This pull-request updates all the absolute jumps and makes them relative
saving a single byte for each change.  Not huge, but why not since we're
already making such changes!