WIP: Emulate BIOS and BDOS calls via OUT instructions
This PR, if merged, would change the way that our emulator handles
syscalls.  The reason for doing this would to increase our accuracy.

At the moment we fall back to the "host" to handle BDOS and BIOS
syscalls when the system IP is a given value - but if other code
is executing at those addresses we get confused.

This came up in the handling of DDT.COM, and A1.COM, and I guess it
is a general problem.

* #112
* #175