Allow input to be scripted (#184)
* Allow input to be scripted

This pull-request allows input to be faked and read from a file.  This is achieved by adding a new console-input driver - using our pluggable system - which just reads from a file.

Using this new facility a simple test-driver was added to perform basic runtime tests of the emulator, using our pre-cooked cpm-dist repository.