Simplify the timeout handling
The Z80 emulator we're using already has support for timeout-handling,
and so we can just use that rather than polling our own ctx.

There's no functional change here - if the code is running inside a
BIOS/BDOS function it is executing on the host and we don't get the
firing of the timeout until control returns to the Z80 - but this is
cleaner.