190 tests (#191)
This pull-request updates the handling of the functional testing, so that we can now have a header in front of each test.  We can now specify per-test options, and we've used that ability to make newline-handling simpler.

The parser for the header/options is now covered by fuzz-testing to ensure there are no obvious issues.

This closes #190. 

We've also fixed linter warnings