More minor test-case updates
This pull-request adds a new configuration option "pause-on-newlines"
which allows automatic delays on newlines, which helps for some of
our test cases.

For example `mbasic2.in` contains:

```
  30 NEXT I
  LIST
  RUN
```

If we let the `LIST` run it will poll for keyboard input to see if
it should react to Ctrl-C, or similar, and that will swallow pending
input.  So we have to pause/pretend there is no input until that
completes.

But adding "`#LIST\n#RUN`" or similar feels fragile and confusing.

Pausing after every newline will make things simpler.