Add the ability to disable our FS and BIOS extension at runtime (#209)
* Add the ability to disable our FS and BIOS extension at runtime

This is essentially a way to allow a configuration to be set
at run-time, via AUTOEXEC.SUB, and then "sealed" such that users
cannot escape and restore configuration - e.g. deny them the
ability to run host-commands.

In practice this is a security-theatre, because anybody who is given
a "locked down cpmulator" can press Ctrl-C and exit to a shell.  Sure
there are scenarious you can imagine with restricted shells, etc, but
in practice this is just a neat thing to be able to do.

This closes #208.

* Added reference to A:\!DISABLE.COM

* Added binary artifact to let tests pass in CI-environment