Change line-breaks for the TYPE-builtin in CCPZ
This PR changes the code in the CCP relating to newline handling
for the TYPE built-in:

* LF -> print "\r" AND "\n", then reset the tab-count.
* CR -> print nothing, but still reset the tab count.

The end result of that should be better output for files that
have only "\n" on their newlines.   However if we ever see a file
that just has "\r" it will be bad.

Ideally this should be configurable "LINE CR", "LINE LF", "LINE CR,LF"
or "LINE LF,CR" for example.

But for the moment this minor change works.