Update to use DMA-area not FCB1 for CLI args.
This pull-request, once complete, will close #213 by changing
the way that the command-line arguments are processed by our
binaries.

Until now we've used the default FCB(1) to look for our arguments
and while this works it suffers from the fact that the CLI argument
is to be coerced into  a filename/pattern, and that has some other
implications such as the fact that some characters are not allowed
and the length is capped.

This PR changes to using the DMA area, which contains the literal
(albeit upper-cased) text the user entered which allows us to enter

   !ccp *
   !ccp some-really-long-value
   !hostcmd a.value.with.multiple.periods

These are not necessary right now, however we want to allow
arbitrary options in the future as part of #212 so doing this
first is a necessary precondition.