Allow input and output drivers to recieve options and use them (#217)
* Allow options to be specified after the driver name.

Allow "name:options" to be specified when our factory method is
invoked.  This will ignore the options when looking for a registered
driver and save the options away - they are not yet used.

* Repeat for consoleout

This will be used for the newline handling in a later update.

* Simple test case

* Test that options don't break driver lookup

* Update allow passing options to !input and !output

This updates to look for options,and to try to deal with them
in our BIOS extensions.  Test cases updated too.

This is a bit messy.  Meh.

* Updated patterns to cope with updated output

* Updated to actually allow control of newlines

* Output a message if only the options have changed

* Updated now we downcase again