Allow changing printer-path at runtime
This pull-request, once complete, will allow changing the printer
logfile at runtime and closes #214.

These syscalls are getting very similar now that we've got our
str-from-userspace and so on.  The only difference in this new
embedded binary is the "OUTPUT_PREFIX", "OUTPUT_SUFFIX" and the
number of the syscall invoked.

Before merging this:

* Unify things, some more.
* Add test-case.