Output shell-commands via our output layer (#220)
We have a notion of pluggable input and output drivers, which are designed to allow us to be flexible and a little portable - though the input layers are more about being functional and avoiding 100% CPU due to tight polling.

The idea of a pluggable output-layer should be that all output is sent via it, however that's not exactly what happens:

    main.go creates a new cpm/ object
        The code in cpm/cpm.go creates an input layer.
        The code in cpm/cpm.go creates an output layer.
    The CCP is almost certainly invoked next, and that results in the cpm/ code processing input
        ReadLine is invoked from the cpm/ package.
        The CCP does its thing, and output is sent to the console

All that sounds great, except for the case where the input layer also writes output. In the command-execution area.

This pull-request updates things so that the shell-execution returns the text output, if any, to the cpm-package where _it_ can send it to the console.

There is still grubbiness as the input-layer runs "printf" on the individual characters input, and handles erases when the user presses backspace/escape.  However logically things are working as they should.

We now have 100% test-coverage of the cpm-package and most of the consolein drivers too.

This closes #219.