Fixed read-test to work correctly
This program was only testing the first byte of each read-record which
meant that it did not accomplish its intended purpose.   Also failures
of that first byte would show a dump of memory which was bogus.

Fixed these, which closes #221.  However I think more could be done
to improve this program.