Allow stuffing input into the console buffer (#228)
This pull-request adds a new CLI argument, so that you can inject
input into the console-buffer from the start:

```
$ cpmulator -cd ../cpm-dist/   -stuff="type /\n" -input stty
```

This is enough to demonstrate the crash reported in #226, without
any annoying blocking.  Now that this is complete it closes #227.

There is a minor complication, if you're setup so that you have
a lot of jobs in "autoexec.sub" then the processing of that submit
file might get killed - processing polls for keyboard input.  Ugh.