Don't abort emulation on file-open error
This pull-request closes #226 by turning a failure to open a file
into something that just results in the BDOS returning an error
code to the calling application - rather than an immediate halt
to the emulation.

This is more akin to what we should do from the start, though it
was such a rare occurence it went unnoticed until I accidentally
entered the bogus "TYPE /" command.