Continue processing SUBMIT-file even if I/O pending
This pull-request is related to the PR we made in #228, which allowed
stuffing console input from the main driver.  The issue we noted was
that we run "SUBMIT AUTOEXEC" on startup, if the appropriate files are
present, and this might get interrupted by the extra stuffed input.

The reason is that the submit-handler in the CCP polls for console
input and stops running if any is round - essentially eating the
input we expected to be read only after the submit file was completed.

This PR changes the CCP to stop polling for console I/O.  This
means "SUBMIT FOO" will write a file "$$$.SUB" and when _that_ file
is read/processed it will NOT be possible to interrupt it.  But that's
 a small price to pay.  This is much more consistent and reliable.