Processing of $$$.SUB files shouldn't poll for console I/O (#230)
* Don't abort emulation on file-open error

This pull-request closes #226 by turning a failure to open a file
into something that just results in the BDOS returning an error
code to the calling application - rather than an immediate halt
to the emulation.

This is more akin to what we should do from the start, though it
was such a rare occurence it went unnoticed until I accidentally
entered the bogus "TYPE /" command.

* Continue processing SUBMIT-file even if I/O pending

This pull-request is related to the PR we made in #228, which allowed
stuffing console input from the main driver.  The issue we noted was
that we run "SUBMIT AUTOEXEC" on startup, if the appropriate files are
present, and this might get interrupted by the extra stuffed input.

The reason is that the submit-handler in the CCP polls for console
input and stops running if any is round - essentially eating the
input we expected to be read only after the submit file was completed.

This PR changes the CCP to stop polling for console I/O.  This
means "SUBMIT FOO" will write a file "$$$.SUB" and when _that_ file
is read/processed it will NOT be possible to interrupt it.  But that's
 a small price to pay.  This is much more consistent and reliable.