Improve implementation after testing with HiSoft C compiler
This pull-request is part of #234, making some improvements but not
yet resolving the issue.   It seems there are some issues with our
implementation which are not yet resolved however there are some
clear issues:

* We're missing some of the syscalls that the binaries call.
* We have issues handling the $-files, as created by SUBMIT.COM

I think we also don't reset FCB details when files are opened which
leads to reading/writing to the wrong offsets if the caller doesn't
pay attention.  (i.e. S1/CR/R0/R1/R2/Ex should be zero-filled when
a file is made/opened, maybe?)