Update to avoid modifying the FCB when file open/made
We've moved to storing the cache of the open files on the host-side
using a key based on the filename.  This means that we no longer
modify the Al[0] and Al[1] members of the FCB in-ram..

This closes #236 by making the caching an internal-detail, rather than
something that guest-binaries could notice (via the changes to Al).