Re-implement file I/O primitives for sanity
This pull-request, once complete, will update #234 by reimplementing
the following CP/M primives:

* File Open
* File Close
* Make File
* Read File
* Write File
* Read Random Record
* Write Random Record

My hope is that this considered and thoughtful rewrite will resolve
the issues plaguing the HiSoft C compiler, however if it does not
then I will close the issue regardless and try to pin down a smaller
and similar way to reproduce the issues found in that binary.

Until this PR is ready for review it will fail all tests and use,
as some/all of the functions have been stubbed out for replacement.